scripts/search/requirements.txt
# Azure Functions dependencies
azure-functions
azurefunctions-extensions-bindings-blob

# Azure SDK dependencies
azure-identity
azure-keyvault-secrets
azure-mgmt-web
azure-mgmt-storage
azure-search-documents==11.5.2
azure-storage-blob
msgraph-sdk==1.5.4

# Data handling and validation dependencies
jsonschema
python-dotenv
requests
openpyxl==3.1.5
tabulate==0.9.0

# AI and NLP dependencies
tiktoken==0.7.0
langchain-text-splitters==0.3.5
openai==1.55.3

# Error handling
tenacity==8.5.0

# Web and async dependencies
aiohttp==3.9.0b0
webvtt-py==0.5.1

# Image library
Pillow==11.1.0

# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
